backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
motor==3.4.0
beanie==1.23.6              # Add for multi-agent features
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
requests==2.32.3
authlib==1.3.0              # Add this (was missing)
starlette==0.38.0
websockets==12.0            # Add for WebSocket support
httpx==0.25.2              # Add for HTTP client
